chore: use u32 for validator counts (#1615)
* chore: use 16 for validator counts

* chore: types.json

* chore: authority set size to use u16

* chore: type alias for AuthorityCount and use in CFE where applicable

chore: missed some

* chore: move db migrations into folder

* chore: key migration using u16 validator count

* chore: migrations

* chore: add remaining AuthorityCount aliases

* chore: migration clean up

* chore: remove unnecessary cast, use AuthorityCount

* chore: inline assignment

* Use u32 authorityCount

* chore: use function to map individual elements

* chore: obey the format

* chore: obey clippy

* chore: allow unit err

* chore: apply translation to new type only the type that changes

+++ state-chain/pallets/cf-auction/src/lib.rs
@@ -207,7 +207,7 @@ impl<T: Config> Auctioneer for Pallet<T> {
 		let (min_number_of_authorities, max_number_of_authorities) =
 			CurrentAuthoritySetSizeRange::<T>::get();
 		// Final rule - Confirm we have our set size
-		ensure!(number_of_bidders >= min_number_of_authorities, {
+		ensure!(number_of_bidders >= min_number_of_authorities as u32, {
 			log::error!(
 				"[cf-auction] insufficient bidders to proceed. {} < {}",
 				number_of_bidders,
@@ -220,7 +220,7 @@ impl<T: Config> Auctioneer for Pallet<T> {
 		bids.reverse();
 
 		let mut target_authority_set_size =
-			min(max_number_of_authorities, number_of_bidders) as usize;
+			min(max_number_of_authorities as u32, number_of_bidders) as usize;
 		let mut next_authority_set: Vec<_> =
 			bids.iter().take(target_authority_set_size as usize).collect();
